Curses UI handles multiple messages
The single line I'm using for messages doesn't handle the fray of multiple attacks all that well. I'm not ready to deal with UI right now, so I'll do a quick hack that makes the user read and acknowledge every message. The nice thing is this becomes a UI issue, not a game issue.